Eror Handeling complete

+++ src/modules/products/products.controller.ts
@@ -1,11 +1,16 @@
 import { Request, Response } from 'express';
 
+import { Product } from './products.model';
 import { ProductServices } from './products.services';
+import { productZodValidationSchema } from './products.validation';
 
 const createProducts = async (req: Request, res: Response) => {
   try {
     const product = req.body;
-    const result = await ProductServices.createProductIntoDB(product);
+    const parseZodValidataionData = productZodValidationSchema.parse(product);
+    const result = await ProductServices.createProductIntoDB(
+      parseZodValidataionData,
+    );
     res.status(200).json({
       success: true,
       message: 'Product created successfully!',
@@ -21,12 +26,29 @@ const createProducts = async (req: Request, res: Response) => {
 };
 const getAllProducts = async (req: Request, res: Response) => {
   try {
-    const result = await ProductServices.getAllProductsFormDB();
-    res.status(200).json({
-      success: true,
-      message: 'Products fetched successfully!',
-      data: result,
-    });
+    const searchTerm = req.query.searchTerm as string | undefined;
+    if (searchTerm) {
+      const result = await ProductServices.searchProductFromDB(searchTerm);
+      if (result.length <= 0) {
+        return res.status(500).json({
+          success: false,
+          message: 'Not Found Product',
+        });
+      }
+      res.status(200).json({
+        success: true,
+        message: `Products matching search term '${searchTerm}' fetched successfully!`,
+        data: result,
+      });
+    } else {
+      const result = await ProductServices.getAllProductsFormDB();
+      res.status(200).json({
+        success: true,
+        message: 'Products fetched successfully!',
+        data: result,
+      });
+    }
+
     // eslint-disable-next-line @typescript-eslint/no-explicit-any
   } catch (error: any) {
     res.status(500).json({
@@ -39,6 +61,14 @@ const getSingleProducts = async (req: Request, res: Response) => {
   try {
     const { productId } = req.params;
 
+    // Check Product id Exists
+    const existsProduct = await Product.isProductIdExits(productId);
+    if (!existsProduct) {
+      return res.status(500).json({
+        success: false,
+        message: 'Product id is not Exists',
+      });
+    }
     const result = await ProductServices.getSingleProductFromDB(productId);
     res.status(200).json({
       success: true,
@@ -53,32 +83,40 @@ const getSingleProducts = async (req: Request, res: Response) => {
     });
   }
 };
-const searchProduct = async (req: Request, res: Response) => {
-  try {
-    const { searchTerm } = req.query;
-    const result = await ProductServices.searchProductFromDB(searchTerm);
-    if (result.length <= 0) {
-      return res.status(500).json({
-        success: false,
-        message: 'Not Found Product',
-      });
-    }
-    res.status(200).json({
-      success: true,
-      message: 'Product fetched successfully!',
-      data: result,
-    });
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  } catch (error: any) {
-    res.status(500).json({
-      success: false,
-      message: error.message || 'Product create False',
-    });
-  }
-};
+// const searchProduct = async (req: Request, res: Response) => {
+//   try {
+//     const { searchTerm } = req.query;
+//     const result = await ProductServices.searchProductFromDB(searchTerm);
+//     if (result.length <= 0) {
+//       return res.status(500).json({
+//         success: false,
+//         message: 'Not Found Product',
+//       });
+//     }
+//     res.status(200).json({
+//       success: true,
+//       message: `Products matching search term ${searchTerm}' fetched successfully!`,
+//       data: result,
+//     });
+//     // eslint-disable-next-line @typescript-eslint/no-explicit-any
+//   } catch (error: any) {
+//     res.status(500).json({
+//       success: false,
+//       message: error.message || 'Product create False',
+//     });
+//   }
+// };
 const deleteProduct = async (req: Request, res: Response) => {
   try {
     const { productId } = req.params;
+    // Check Product id Exists
+    const existsProduct = await Product.isProductIdExits(productId);
+    if (!existsProduct) {
+      return res.status(500).json({
+        success: false,
+        message: 'Product id is not Exists',
+      });
+    }
 
     await ProductServices.deletePrductFormDB(productId);
     res.status(200).json({
@@ -98,6 +136,14 @@ const updateProduct = async (req: Request, res: Response) => {
   try {
     const { productId } = req.params;
     const updatedData = req.body;
+    // Check Product id Exists
+    const existsProduct = await Product.isProductIdExits(productId);
+    if (!existsProduct) {
+      return res.status(500).json({
+        success: false,
+        message: 'Product id is not Exists',
+      });
+    }
 
     const result = await ProductServices.updateProductFormDB(
       productId,
@@ -121,7 +167,7 @@ export const ProductControllers = {
   createProducts,
   getAllProducts,
   getSingleProducts,
-  searchProduct,
+
   deleteProduct,
   updateProduct,
 };
